\name{Kikwit95}
\alias{Kikwit95}
\docType{data}
\title{
    Data for the 1995 Ebola Oubreak in Kikwit, DRC
}
\description{
    Transcribed from Lekone and Finkenstadt, 2006
}
\usage{data("Kikwit95")}
\format{
  A data frame with 197 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{Count}}{a numeric vector}
  }
}
\references{
    Statistical Inference in a Stochastic Epidemic SEIR Model with Control Intervention: Ebola as a Case Study, Lekone and Finkenstadt, 2006.
}
\examples{
data(Kikwit95)
}
\keyword{datasets}
