\name{KikwitFullSpecResults}
\alias{KikwitFullSpecResults}
\docType{data}
\title{
    Analysis output for the 1995 Ebola Outbreak in Kikwit, DRC. 
}
\description{  
    A list containing raw and Processed MCMC results for the fully specified analysis of the Kikwit Ebola outbreak. 
}
\keyword{datasets}
