\name{KikwitUnderspecResults}
\alias{KikwitUnderspecResults}
\docType{data}
\title{
    Analysis output for the 1995 Ebola Outbreak in Kikwit, DRC. 
}
\description{  
    A list containing raw and Processed MCMC results for the underspecified analysis of the Kikwit Ebola outbreak. 
}
\keyword{datasets}
