\name{WestAfrica0DFResults}
\alias{WestAfrica0DFResults}
\docType{data}
\title{
    MCMC output for the 2014-2015 Ebola Outbreak in West Africa
}
\description{
    Raw and processed MCMC samples for the intercept only spatial SEIR analysis of the 2014-2015 Ebola outbreak in West Africa.
}
\keyword{datasets}
