\name{WestAfrica5DFResults}
\alias{WestAfrica5DFResults}
\docType{data}
\title{
    MCMC output for the 2014-2015 Ebola Outbreak in West Africa
}
\description{
    Raw and processed MCMC samples for the intercept + 5 DF temporal basis spatial SEIR analysis of the 2014-2015 Ebola outbreak in West Africa.
}
\keyword{datasets}
